Fix for 64-bit build in Windows 10: pointer arithmetic result now held in a 64-bit variable.



See merge request !104

+++ test/support_tests.c
@@ -90,12 +90,12 @@ void initFullResourcePath(const char* rootDir) {
 }
 
 static void escapeStr(const char *str, CuString *escaped) {
-	int p;
+	long long int p;
 	static const char *replIndex = "<>&\"'";
 	static const char *repl[] = { "lt", "gt", "amp", "quot", "#39"};
 	while (*str) {
 		/* Find the index of current char. */
-		p = (int)(strchr(replIndex, *str) - replIndex);
+		p = (long long int)(strchr(replIndex, *str) - replIndex);
 		/* If the character is found, use the replacement */
 		if (p >= 0) {
 			CuStringAppendFormat(escaped, "&%s", repl[p]);
